requirements.txt
asgiref==3.11.0
crispy-bootstrap4==2025.6
crispy-bootstrap5==2025.6
Django==6.0
django-background-tasks==1.2.8
django-crispy-forms==2.5
django-crontab==0.7.1
django-widget-tweaks==1.5.0
pillow==12.0.0
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
